feat: Relocate media type
